code/deployment/requirements.txt
azure-eventhub==5.5.0
azure-eventhub-checkpointstoreblob-aio==1.1.4
azure-identity==1.2.0
azure-keyvault==1.1.0
azure-keyvault-secrets==4.0.1
azure-schemaregistry==1.0.0b1
azure-servicebus==0.50.3
azure-storage-blob==12.6.0
azure-storage-file-datalake==12.3.1
avro==1.10.2
opencensus-ext-azure==1.0.2
pandas==0.25.3  
pyodbc==4.0.30
